refactor(JPQL) ISS-010: Adjust to get from repository
