+++ src/main/java/com/algaworks/algafood/jpa/BuscaCozinhaMain.java
@@ -2,6 +2,7 @@ package com.algaworks.algafood.jpa;
 
 import com.algaworks.algafood.AlgafoodApiApplication;
 import com.algaworks.algafood.domain.model.Cozinha;
+import com.algaworks.algafood.domain.repository.CozinhaRepository;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.context.ConfigurableApplicationContext;
@@ -15,10 +16,10 @@ public class BuscaCozinhaMain {
                 .run(args);
 
         // Pegamos o Bean que inicializamos no Cadastro cozinha para podermos realizar as consultas
-        CadastroCozinha cadastroCozinha = applicationContext.getBean(CadastroCozinha.class);
+        CozinhaRepository cozinhaRepository = applicationContext.getBean(CozinhaRepository.class);
 
         // Buscando cozinha com id 1
-        Cozinha cozinha = cadastroCozinha.buscar(1L);
+        Cozinha cozinha = cozinhaRepository.buscarPorId(1L);
 
         System.out.println(cozinha.getNome());
