+++ src/main/java/com/algaworks/algafood/jpa/ExclusaoCozinhaMain.java
@@ -2,6 +2,7 @@ package com.algaworks.algafood.jpa;
 
 import com.algaworks.algafood.AlgafoodApiApplication;
 import com.algaworks.algafood.domain.model.Cozinha;
+import com.algaworks.algafood.domain.repository.CozinhaRepository;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.context.ConfigurableApplicationContext;
@@ -15,12 +16,12 @@ public class ExclusaoCozinhaMain {
                 .run(args);
 
         // Pegamos o Bean que inicializamos no Cadastro cozinha para podermos realizar as consultas
-        CadastroCozinha cadastroCozinha = applicationContext.getBean(CadastroCozinha.class);
+        CozinhaRepository cozinhaRepository = applicationContext.getBean(CozinhaRepository.class);
 
         Cozinha cozinha = new Cozinha();
         cozinha.setId(1L);
 
-        cadastroCozinha.remover(cozinha);
+        cozinhaRepository.remover(cozinha);
 
     }
