+++ src/main/java/com/algaworks/algafood/jpa/InclusaoCozinhaMain.java
@@ -2,6 +2,7 @@ package com.algaworks.algafood.jpa;
 
 import com.algaworks.algafood.AlgafoodApiApplication;
 import com.algaworks.algafood.domain.model.Cozinha;
+import com.algaworks.algafood.domain.repository.CozinhaRepository;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.context.ConfigurableApplicationContext;
@@ -15,7 +16,7 @@ public class InclusaoCozinhaMain {
                 .run(args);
 
         // Pegamos o Bean que inicializamos no Cadastro cozinha para podermos realizar as consultas
-        CadastroCozinha cadastroCozinha = applicationContext.getBean(CadastroCozinha.class);
+        CozinhaRepository cozinhaRepository = applicationContext.getBean(CozinhaRepository.class);
 
         Cozinha cozinha1 = new Cozinha();
         cozinha1.setNome("Brasileira");
@@ -24,8 +25,8 @@ public class InclusaoCozinhaMain {
         cozinha2.setNome("Japonesa");
 
         // Nesse ponto do código os Id's ainda estão null, após adicionar, irão autoincrementar.
-        cadastroCozinha.salvar(cozinha1);
-        cadastroCozinha.salvar(cozinha2);
+        cozinhaRepository.salvar(cozinha1);
+        cozinhaRepository.salvar(cozinha2);
 
         // Provando que os Id's de cozinha 1 e 2 estão null, pois estamos pegando da variável que instanciamos inicialmente
         System.out.printf("%d - %s\n", cozinha1.getId(), cozinha1.getNome());
